+++ src/hooks/fetch/useSummaryQuery.ts
@@ -1,6 +1,7 @@
-import { getSummary } from '@/services/summary'
 import { useQuery } from '@tanstack/react-query'
 import { useParams } from 'next/navigation'
+
+import { getSummary } from '@/services/summary/api'
 
 export function useSummaryQuery() {
   const params = useParams()
@@ -11,6 +12,7 @@ export function useSummaryQuery() {
     queryFn: () => getSummary(`${id}`),
     refetchInterval: data => {
       const status = data?.state?.data?.data?.data?.status
+
       return status === 'completed' || status === 'failed' ? false : 1000
     },
     enabled: !!id,
